pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-intel-tool"
description = "A simple parser for D-Scans and more for Alliance Auth"
readme = "README.md"
keywords = [
    "allianceauth",
    "eveonline",
]
license = {file = "LICENSE"}
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.8"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.0.0b1",
    "allianceauth-app-utils>=1.19.1",
    "django-eveuniverse>=1.3",
]
[project.optional-dependencies]
tests-allianceauth-latest = [
    "coverage",
    "django-webtest",
]
[project.urls]
Changelog = "https://github.com/ppfeufer/aa-intel-tool/blob/master/CHANGELOG.md"
Documentation = "https://github.com/ppfeufer/aa-intel-tool/blob/master/README.md"
Donations = "https://ko-fi.com/ppfeufer"
Homepage = "https://github.com/ppfeufer/aa-intel-tool"
Source = "https://github.com/ppfeufer/aa-intel-tool.git"
Tracker = "https://github.com/ppfeufer/aa-intel-tool/issues"

[tool.hatch.build]
include = [
    "/aa_intel_tool",
]

[tool.hatch.version]
path = "aa_intel_tool/__init__.py"
